refactor: move compareValues util function to separate folder
